LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.arcTo.nonfinite.worker-expected.txt


PASS arcTo() with Infinity/NaN is ignored 
